Document public interface.
Rename send_frame to sendFrame.

Deprecate send_frame.

Add onclose, onframe, onconnectedframe, onmessageframe, onreceiptframe and onerrorframe callbacks.

Incompatible change: the onopen callback is now called when the transport (socket) is openned and not when the STOMP session is succefully established.  you should use the "onconnectedframe" as a replacement for the old behaviour.